refactor(tool): Move template-related operations
Move files for template-related operations to `template/` directory to
allow the helper tool have two types of operations (template-related and
fingerprint-related operations).
