+++ src/cpu-template-helper/src/template/mod.rs
@@ -1,0 +1,6 @@
+// Copyright 2023 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+pub mod dump;
+pub mod strip;
+pub mod verify;
