+++ src/cpu-template-helper/src/template/strip/mod.rs
