+++ src/cpu-template-helper/src/template/strip/x86_64.rs
@@ -4,7 +4,7 @@
 use vmm::cpu_config::templates::CustomCpuTemplate;
 use vmm::cpu_config::x86_64::custom_cpu_template::{CpuidLeafModifier, RegisterModifier};
 
-use crate::strip::strip_common;
+use crate::template::strip::strip_common;
 use crate::utils::x86_64::{CpuidModifierMap, MsrModifierMap};
 
 #[allow(dead_code)]
